requirements.txt
llama-index==0.12.11
llama-index-agent-openai==0.4.2
llama-index-cli==0.4.0
llama-index-core==0.12.11
llama-index-embeddings-huggingface==0.5.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-llms-openai==0.3.13
llama-index-multi-modal-llms-openai==0.4.2
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.3
llama-index-readers-github==0.5.0
llama-index-readers-llama-parse==0.4.0
snowflake-sqlalchemy==1.7.3
trulens-connectors-snowflake==1.3.0
trulens-core==1.3.0
trulens-providers-cortex==1.3.0
